Two options available: Static and Dynamic Placement. Working on interface

+++ RectPacking/Models/IRated.cs
@@ -8,6 +8,6 @@ namespace RectPacking.Models
 {
     public interface IRated
     {
-        int Rating { get; set; }
+        double Rating { get; set; }
     }
 }
